+++ src/OmniSharp.Abstractions/Models/Events/ProjectConfigurationMessage.cs
@@ -8,6 +8,7 @@ namespace OmniSharp.Models.Events
         public string ProjectId { get; set; }
         public string SessionId { get; set; }
         public int OutputKind { get; set; }
+        public IEnumerable<string> ProjectCapabilities { get; set; }
         public IEnumerable<string> TargetFrameworks { get; set; }
         public string SdkVersion { get; set; }
         public IEnumerable<string> References { get; set; }