+++ src/OmniSharp.MSBuild/ProjectFile/ItemNames.cs
@@ -8,6 +8,7 @@ namespace OmniSharp.MSBuild.ProjectFile
         public const string Content = nameof(Content);
         public const string EditorConfigFiles = nameof(EditorConfigFiles);
         public const string PackageReference = nameof(PackageReference);
+        public const string ProjectCapability = nameof(ProjectCapability);
         public const string ProjectReference = nameof(ProjectReference);
         public const string ReferencePath = nameof(ReferencePath);
     }